update import dependencies
