+++ scripts/bert/tokenization.py
@@ -21,7 +21,7 @@ from __future__ import print_function
 import collections
 import unicodedata
 import six
-import tensorflow as tf
+import io
 
 
 def convert_to_unicode(text):
@@ -45,7 +45,7 @@ def convert_to_unicode(text):
 
 
 def printable_text(text):
-  """Returns text encoded in a way suitable for print or `tf.logging`."""
+  """Returns text encoded in a way suitable for print."""
 
   # These functions want `str` for both Python2 and Python3, but in one case
   # it's a Unicode string and in the other it's a byte string.
@@ -71,7 +71,7 @@ def load_vocab(vocab_file):
   """Loads a vocabulary file into a dictionary."""
   vocab = collections.OrderedDict()
   index = 0
-  with tf.gfile.GFile(vocab_file, "r") as reader:
+  with io.open(vocab_file, "r") as reader:
     while True:
       token = convert_to_unicode(reader.readline())
       if not token:
